requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
psycopg2==2.9.10
PyJWT==2.9.0
requests==2.32.3
setuptools==75.2.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
